requirements.txt
aio-timers==0.0.1
aiohttp==3.8.1
async-timeout==4.0.2
black==21.12b0
emoji==1.2.0
google==3.0.0
nextcord==2.0.0a8
psycopg2-binary==2.9.1
python-dotenv==0.16.0
regex==2021.8.28
SQLAlchemy==1.4.25
typing-extensions==4.0.1
urllib3==1.26.4
zipp==3.4.1
